Dashboards: Remove legacy searcher
