+++ pkg/registry/apis/dashboard/search_test.go
@@ -3,7 +3,6 @@ package dashboard
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http/httptest"
 	"net/url"
 	"testing"
@@ -27,153 +26,15 @@ import (
 	"github.com/grafana/grafana/pkg/storage/unified/resourcepb"
 )
 
-func TestSearchFallback(t *testing.T) {
-	t.Run("should hit legacy search handler on mode 0", func(t *testing.T) {
+func TestSearch(t *testing.T) {
+	t.Run("should hit unified storage search handler", func(t *testing.T) {
 		mockClient := &MockClient{}
-		mockLegacyClient := &MockClient{}
-
-		cfg := &setting.Cfg{
-			UnifiedStorage: map[string]setting.UnifiedStorageConfig{
-				"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode0},
-			},
-		}
-		dual := dualwrite.ProvideStaticServiceForTests(cfg)
-		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockLegacyClient, mockClient, nil)
-
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest("GET", "/search", nil)
-		req.Header.Add("content-type", "application/json")
-		req = req.WithContext(identity.WithRequester(req.Context(), &user.SignedInUser{Namespace: "test"}))
-
-		searchHandler.DoSearch(rr, req)
-
-		if mockClient.LastSearchRequest != nil {
-			t.Fatalf("expected Search NOT to be called, but it was")
-		}
-		if mockLegacyClient.LastSearchRequest == nil {
-			t.Fatalf("expected Search to be called, but it was not")
-		}
-	})
-
-	t.Run("should hit legacy search handler on mode 1", func(t *testing.T) {
-		mockClient := &MockClient{}
-		mockLegacyClient := &MockClient{}
-
-		cfg := &setting.Cfg{
-			UnifiedStorage: map[string]setting.UnifiedStorageConfig{
-				"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode1},
-			},
-		}
-		dual := dualwrite.ProvideStaticServiceForTests(cfg)
-		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockLegacyClient, mockClient, nil)
-
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest("GET", "/search", nil)
-		req.Header.Add("content-type", "application/json")
-		req = req.WithContext(identity.WithRequester(req.Context(), &user.SignedInUser{Namespace: "test"}))
-
-		searchHandler.DoSearch(rr, req)
-
-		if mockClient.LastSearchRequest != nil {
-			t.Fatalf("expected Search NOT to be called, but it was")
-		}
-		if mockLegacyClient.LastSearchRequest == nil {
-			t.Fatalf("expected Search to be called, but it was not")
-		}
-	})
-
-	t.Run("should hit legacy search handler on mode 2", func(t *testing.T) {
-		mockClient := &MockClient{}
-		mockLegacyClient := &MockClient{}
-
-		cfg := &setting.Cfg{
-			UnifiedStorage: map[string]setting.UnifiedStorageConfig{
-				"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode2},
-			},
-		}
-		dual := dualwrite.ProvideStaticServiceForTests(cfg)
-		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockLegacyClient, mockClient, nil)
-
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest("GET", "/search", nil)
-		req.Header.Add("content-type", "application/json")
-		req = req.WithContext(identity.WithRequester(req.Context(), &user.SignedInUser{Namespace: "test"}))
-
-		searchHandler.DoSearch(rr, req)
-
-		if mockClient.LastSearchRequest != nil {
-			t.Fatalf("expected Search NOT to be called, but it was")
-		}
-		if mockLegacyClient.LastSearchRequest == nil {
-			t.Fatalf("expected Search to be called, but it was not")
-		}
-	})
-
-	t.Run("should hit legacy search handler on mode 3", func(t *testing.T) {
-		mockClient := &MockClient{}
-		mockLegacyClient := &MockClient{}
-
-		cfg := &setting.Cfg{
-			UnifiedStorage: map[string]setting.UnifiedStorageConfig{
-				"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode3},
-			},
-		}
-		dual := dualwrite.ProvideStaticServiceForTests(cfg)
-		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockLegacyClient, mockClient, nil)
-
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest("GET", "/search", nil)
-		req.Header.Add("content-type", "application/json")
-		req = req.WithContext(identity.WithRequester(req.Context(), &user.SignedInUser{Namespace: "test"}))
-
-		searchHandler.DoSearch(rr, req)
-
-		if mockClient.LastSearchRequest != nil {
-			t.Fatalf("expected Search NOT to be called, but it was")
-		}
-		if mockLegacyClient.LastSearchRequest == nil {
-			t.Fatalf("expected Search to be called, but it was not")
-		}
-	})
-
-	t.Run("should hit unified storage search handler on mode 4", func(t *testing.T) {
-		mockClient := &MockClient{}
-		mockLegacyClient := &MockClient{}
-
-		cfg := &setting.Cfg{
-			UnifiedStorage: map[string]setting.UnifiedStorageConfig{
-				"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode4},
-			},
-		}
-		dual := dualwrite.ProvideStaticServiceForTests(cfg)
-		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockLegacyClient, mockClient, nil)
-
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest("GET", "/search", nil)
-		req.Header.Add("content-type", "application/json")
-		req = req.WithContext(identity.WithRequester(req.Context(), &user.SignedInUser{Namespace: "test"}))
-
-		searchHandler.DoSearch(rr, req)
-
-		if mockClient.LastSearchRequest == nil {
-			t.Fatalf("expected Search to be called, but it was not")
-		}
-		if mockLegacyClient.LastSearchRequest != nil {
-			t.Fatalf("expected Search NOT to be called, but it was")
-		}
-	})
-
-	t.Run("should hit unified storage search handler on mode 5", func(t *testing.T) {
-		mockClient := &MockClient{}
-		mockLegacyClient := &MockClient{}
-
-		cfg := &setting.Cfg{
+		dual := dualwrite.ProvideStaticServiceForTests(&setting.Cfg{
 			UnifiedStorage: map[string]setting.UnifiedStorageConfig{
 				"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode5},
 			},
-		}
-		dual := dualwrite.ProvideStaticServiceForTests(cfg)
-		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockLegacyClient, mockClient, nil)
+		})
+		searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockClient, nil)
 
 		rr := httptest.NewRecorder()
 		req := httptest.NewRequest("GET", "/search", nil)
@@ -184,101 +45,6 @@ func TestSearchFallback(t *testing.T) {
 
 		if mockClient.LastSearchRequest == nil {
 			t.Fatalf("expected Search to be called, but it was not")
-		}
-		if mockLegacyClient.LastSearchRequest != nil {
-			t.Fatalf("expected Search NOT to be called, but it was")
-		}
-	})
-}
-
-func TestSearchHandlerPagination(t *testing.T) {
-	t.Run("should calculate offset and page parameters", func(t *testing.T) {
-		limit := 50
-		for i, tt := range []struct {
-			offset         int
-			page           int
-			expectedOffset int
-			expectedPage   int
-		}{
-			{
-				offset:         0,
-				page:           0,
-				expectedOffset: 0,
-				expectedPage:   1,
-			},
-			{
-				offset:         0,
-				page:           1,
-				expectedOffset: 0,
-				expectedPage:   1,
-			},
-			{
-				offset:         0,
-				page:           2,
-				expectedOffset: 50,
-				expectedPage:   2,
-			},
-			{
-				offset:         0,
-				page:           3,
-				expectedOffset: 100,
-				expectedPage:   3,
-			},
-			{
-				offset:         50,
-				page:           0,
-				expectedOffset: 50,
-				expectedPage:   2,
-			},
-			{
-				offset:         100,
-				page:           0,
-				expectedOffset: 100,
-				expectedPage:   3,
-			},
-			{
-				offset:         149,
-				page:           0,
-				expectedOffset: 149,
-				expectedPage:   3,
-			},
-			{
-				offset:         150,
-				page:           0,
-				expectedOffset: 150,
-				expectedPage:   4,
-			},
-		} {
-			mockClient := &MockClient{}
-
-			cfg := &setting.Cfg{
-				UnifiedStorage: map[string]setting.UnifiedStorageConfig{
-					"dashboards.dashboard.grafana.app": {DualWriterMode: rest.Mode0},
-				},
-			}
-			dual := dualwrite.ProvideStaticServiceForTests(cfg)
-			searchHandler := NewSearchHandler(tracing.NewNoopTracerService(), dual, mockClient, mockClient, nil)
-
-			rr := httptest.NewRecorder()
-			endpoint := fmt.Sprintf("/search?limit=%d", limit)
-			if tt.offset > 0 {
-				endpoint = fmt.Sprintf("%s&offset=%d", endpoint, tt.offset)
-			}
-			if tt.page > 0 {
-				endpoint = fmt.Sprintf("%s&page=%d", endpoint, tt.page)
-			}
-			req := httptest.NewRequest("GET", endpoint, nil)
-			req.Header.Add("content-type", "application/json")
-			req = req.WithContext(identity.WithRequester(req.Context(), &user.SignedInUser{Namespace: "test"}))
-
-			searchHandler.DoSearch(rr, req)
-
-			if mockClient.LastSearchRequest == nil {
-				t.Fatalf("expected Search to be called, but it was not")
-			}
-
-			require.Equal(t, int(mockClient.LastSearchRequest.Offset), tt.expectedOffset, fmt.Sprintf("mismatch offset in test %d", i))
-			require.Equal(t, int(mockClient.LastSearchRequest.Page), tt.expectedPage, fmt.Sprintf("mismatch page in test %d", i))
 		}
 	})
 }
